Developed initialization code
Created a file, bench_initializers.cpp, with which
we can compare the performance (time and msse) on initializers.
An example of running this file is in init.sh.

Also created an LSH initialization method (hash the points, sort them,
chunk them into even block sizes, and find a center for each block).
It works pretty well, gets a worse msse than Macqueen but x100 runtime
(depending on input) (it is possible that this is an implementation flaw
in Macqueen). Please note the hashing and hyperplane generation
is essentially the same as the one in grann.