Experiment 17 results
This experiment is a third run on initialization benches,
to get graphs for the presentation.

This version has two improvements:
first, we run k=10,100,200,400,500 so we can show on our graph
kmeans++ getting way too big.
second, we run a version of Forgy's, called ForgyRandom, that
actually picks random assignments, instead of using
asg[i] = i % k (not truly random) like the previous implementation,
which has the potential to perform better or worse if there is some
pattern in the data.