Improving yy code
From the benching against eakmeans, it seems that our yy
code can be notably improved. This commit is the start of
that process. In include/yy_improved I've started reviewing code
and optimizing where possible. The code is buggy currently --
use the previous Yinyang code for runs for the time being.