Start of PQ kmeans code
Starting building the PQ kmeans code, which uses product quantization to reduce the time that the distance calculations take.
On the way to doing this, wrote a cleaner and shorter version of naive k-means, in naive2.h

Bug concern: in compute centers, the c array (where we do the adding) should be set to 0? not the previous center value.
We only set to the previous center value if the center has no members. Also the yy code has large variance of runtime
-- ought to do some correctness testing across the board. Currently the naive2 and yy_improved code both give realistic
results but the msse's are more different than normal (77K vs 78K) which may or may not be an issue.